feat: expose internal ComponentList component for direct rendering of changelog data
