+++ lib/components/ChangelogList/MinimalChangelogList.tsx
@@ -2,13 +2,11 @@ import * as React from 'react';
 import { useUpdateHiveContext } from '../ChangelogContext';
 import {
   ChangelogWithComponents,
-  ChangeTypeMap,
-  getTypeColor,
   ungroupedChangelogs,
 } from '../changelog.util.ts';
 import { ChangeType } from '../../changelog.types.ts';
 import { useMemo } from 'react';
-import ComponentList from './_internal/ComponentList.tsx';
+import { ComponentList } from './ComponentList.tsx';
 
 interface Props {
   changeTypeMapper?: Record<ChangeType, string>;
@@ -21,9 +19,7 @@ interface Props {
  *
  * @param changeTypeMapper Overridable mapping of change types to displayable representations.
  */
-export const MinimalChangelogList: React.FC<Props> = ({
-  changeTypeMapper = ChangeTypeMap,
-}) => {
+export const MinimalChangelogList: React.FC<Props> = ({ changeTypeMapper }) => {
   const { data } = useUpdateHiveContext();
 
   const componentChangelogs: ChangelogWithComponents[] | undefined =
@@ -41,7 +37,6 @@ export const MinimalChangelogList: React.FC<Props> = ({
         <ComponentList
           changelogs={componentChangelogs}
           changeTypeMapper={changeTypeMapper}
-          typeColorResolver={getTypeColor}
           hideEntryType
         />
       )}
