+++ lib/components/ChangelogList/index.ts
@@ -2,3 +2,4 @@ export { GroupBy } from './ChangelogList.types';
 
 export { MinimalChangelogList } from './MinimalChangelogList';
 export { ChangelogList } from './ChangelogList';
+export { ComponentList } from './ComponentList';
